requirements-dev.txt
-r requirements.txt
astroid==3.3.5
dill==0.3.9
django-debug-toolbar==4.4.6
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
packaging==24.1
platformdirs==4.3.6
pluggy==1.5.0
pylint==3.3.1
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
pytest==8.3.3
tomlkit==0.13.2
flake8==7.1.1
pyflakes==3.2.0
black==24.10.0